[Modules][ImGui] Corrected index buffer offset calculation in ImGui::Context module
This commit improves the calculation of the index buffer offset within the ImGui::Context module. The previous version erroneously incremented the offset by the size of the vertex buffer for each command in the draw list. This misalignment of indices might have caused incorrect rendering or graphical artifacts. The fix ensures that the correct offset is used for each index.