[Core][Graphics][Metal] Fixed argument descriptor corruption
There is apparently a bug in Metal when enabling GPU validation, where calls to MTL::Device::newArgumentEncoder() can have side effects on the passed data, leading to data corruption. This lead to issues further down the line.
This was fixed by performing a copy of the argument descriptors before passing them to the method.
Apple has been notified of the issue.